# docs/swagger.yaml
basePath: /
definitions:
...
  database.CreateShipmentParams:
    properties:
      amount:
        items:
          type: integer
        type: array
      d_location:
        $ref: '#/definitions/util.LocationParams'
      d_service:
        type: integer
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      p_location:
        $ref: '#/definitions/util.LocationParams'
      p_service:
        type: integer
      priority:
        example: 10
        type: integer
      skills:
        example:
        - 1
        - 5
        items:
          type: integer
        type: array
    required:
    - d_location
    - p_location
    type: object
  database.CreateShipmentTimeWindowParams:
    properties:
      kind:
        type: string
      tw_close:
        example: "2021-12-31 23:59:00"
        type: string
      tw_open:
        example: "2021-12-31 23:00:00"
        type: string
    required:
    - kind
    - tw_close
    - tw_open
    type: object
...
  database.Shipment:
    properties:
      amount:
        items:
          type: integer
        type: array
      created_at:
        example: "2021-12-01 13:00:00"
        type: string
      d_location:
        $ref: '#/definitions/util.LocationParams'
      d_service:
        type: integer
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      id:
        example: "1234567812345678"
        type: string
      p_location:
        $ref: '#/definitions/util.LocationParams'
      p_service:
        type: integer
      priority:
        example: 10
        type: integer
      project_id:
        example: "1234567812345678"
        type: string
      skills:
        example:
        - 1
        - 5
        items:
          type: integer
        type: array
      updated_at:
        example: "2021-12-01 13:00:00"
        type: string
    type: object
  database.ShipmentTimeWindow:
    properties:
      created_at:
        example: "2021-12-01 13:00:00"
        type: string
      id:
        example: "1234567812345678"
        type: string
      kind:
        type: string
      tw_close:
        example: "2021-12-31 23:59:00"
        type: string
      tw_open:
        example: "2021-12-31 23:00:00"
        type: string
      updated_at:
        example: "2021-12-01 13:00:00"
        type: string
    type: object
  database.UpdateShipmentParams:
    properties:
      amount:
        items:
          type: integer
        type: array
      d_location:
        $ref: '#/definitions/util.LocationParams'
      d_service:
        type: integer
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      p_location:
        $ref: '#/definitions/util.LocationParams'
      p_service:
        type: integer
      priority:
        example: 10
        type: integer
      project_id:
        example: "1234567812345678"
        type: string
      skills:
        example:
        - 1
        - 5
        items:
          type: integer
        type: array
    type: object
...
  util.LocationParams:
    properties:
      latitude:
        example: 2.0365
        type: number
      longitude:
        example: 48.6113
        type: number
    required:
    - latitude
    - longitude
    type: object
  util.MultiError:
    properties:
      errors:
        example:
        - Error message1
        - Error message2
        items:
          type: string
        type: array
    type: object
  util.NotFound:
    properties:
      error:
        example: Not Found
        type: string
    type: object
...
  util.Success:
    properties:
      success:
        example: "true"
        type: string
    type: object
host: localhost:9100
info:
  contact:
    email: info@georepublic.de
    name: Team Georepublic
  description: This is an API for scheduling VRP tasks. Source code can be found on
    https://github.com/Georepublic/pg_scheduleserv
  license:
    name: GNU Affero General Public License
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://swagger.io/terms/
  title: pg_scheduleserv API
  version: 0.1.0
paths:
...
  /projects/{project_id}/shipments:
    get:
      consumes:
      - application/json
      description: Get a list of shipments for a project with project_id
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: List shipments for a project
      tags:
      - Shipment
    post:
      consumes:
      - application/json
      description: Create a new shipment with the input payload
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Create shipment
        in: body
        name: Shipment
        required: true
        schema:
          $ref: '#/definitions/database.CreateShipmentParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Create a new shipment
      tags:
      - Shipment
...
  /shipments/{shipment_id}:
    delete:
      consumes:
      - application/json
      description: Delete a shipment with its shipment_id
      parameters:
      - description: Shipment ID
        in: path
        name: shipment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Delete a shipment
      tags:
      - Shipment
    get:
      consumes:
      - application/json
      description: Fetch a shipment with its shipment_id
      parameters:
      - description: Shipment ID
        in: path
        name: shipment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.NotFound'
      summary: Fetch a shipment
      tags:
      - Shipment
    patch:
      consumes:
      - application/json
      description: Update a shipment with its shipment_id
      parameters:
      - description: Shipment ID
        in: path
        name: shipment_id
        required: true
        type: integer
      - description: Update shipment
        in: body
        name: Shipment
        required: true
        schema:
          $ref: '#/definitions/database.UpdateShipmentParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Update a shipment
      tags:
      - Shipment
  /shipments/{shipment_id}/time_windows:
    get:
      consumes:
      - application/json
      description: Get a list of shipment time windows for a shipment with shipment_id
      parameters:
      - description: Shipment ID
        in: path
        name: shipment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.ShipmentTimeWindow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: List shipment time windows for a shipment
      tags:
      - Shipment
    post:
      consumes:
      - application/json
      description: Create a new shipment time window with the input payload
      parameters:
      - description: Shipment ID
        in: path
        name: shipment_id
        required: true
        type: integer
      - description: Create shipment time window
        in: body
        name: ShipmentTimeWindow
        required: true
        schema:
          $ref: '#/definitions/database.CreateShipmentTimeWindowParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.ShipmentTimeWindow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Create a new shipment time window
      tags:
      - Shipment
  /shipments/{shipment_id}/time_windows/{tw_open}/{tw_close}:
    delete:
      consumes:
      - application/json
      description: Delete shipment time windows for a shipment with shipment_id
      parameters:
      - description: Shipment ID
        in: path
        name: shipment_id
        required: true
        type: integer
      - description: Shipment opening Time Window
        in: path
        name: tw_open
        required: true
        type: string
      - description: Shipment closing Time Window
        in: path
        name: tw_close
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Delete shipment time windows
      tags:
      - Shipment
...
schemes:
- http
- https
swagger: "2.0"